% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freightdataus-package.R
\docType{package}
\name{freightdataus-package}
\alias{freightdataus}
\alias{freightdataus-package}
\title{freightdataus: Interregional Freight Flow Data for the United States}
\description{
Provides US freight flow data from the Commodity Flow Survey (CFS) Public Use File for 2012, 2017, and 2022: shipment records, weighted origin-destination flows by commodity and mode, zones, and distances between zones. The data are built on demand from files published by the US Census Bureau using the 'tarchives' and 'targets' packages, and are returned in the same format for every survey year.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/freightdataus}
  \item \url{https://uchidamizuki.github.io/freightdataus/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/freightdataus/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

Authors:
\itemize{
  \item Mizuki Uchida \email{uchidamizuki@vivaldi.net}
}

}
\keyword{internal}
